+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/utils/extensions/UpdatesPoller.kt
@@ -31,7 +31,8 @@ class UpdatesPoller(
     private suspend fun pushMediaGroupUpdate(update: BaseMessageUpdate? = null) {
         val inputMediaGroupId = (update ?.data as? MediaGroupMessage) ?.mediaGroupId
         if (mediaGroup.isNotEmpty() && inputMediaGroupId ?.equals(mediaGroup.mediaGroupId) != true) {
-            listOf(*mediaGroup.toTypedArray()).toMediaGroupUpdate() ?.let {
+            mediaGroup.sortBy { it.updateId }
+            mediaGroup.toMediaGroupUpdate() ?.let {
                 sendToBlock(it)
             } ?: mediaGroup.forEach {
                 sendToBlock(it)