+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/CodeTextMessageEntity.kt
@@ -5,5 +5,6 @@ data class CodeTextMessageEntity(
     override val length: Int,
     override val sourceString: String
 ) : TextMessageEntity() {
-    override val formatSymbol: String = "`"
+    override val markdownFormatSymbol: String = "`"
+    override val htmlFormatTagname: String = "code"
 }