+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/EMailMessageEntity.kt
@@ -4,8 +4,7 @@ class EMailMessageEntity(
     override val offset: Int,
     override val length: Int,
     override val sourceString: String
-) : MessageEntity {
-    override val asMarkdownSource: String by lazy {
-        "[$sourceString](mailto://$sourceString)"
-    }
-}
+) : LinkMessageEntity(
+    sourceString,
+    "mailto://$sourceString"
+)