+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/MentionMessageEntity.kt
@@ -4,7 +4,4 @@ class MentionMessageEntity(
     override val offset: Int,
     override val length: Int,
     override val sourceString: String
-) : MessageEntity {
-    override val asMarkdownSource: String
-        get() = sourceString
-}
+) : MessageEntity