+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/MessageEntity.kt
@@ -1,5 +1,6 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity
 
+import com.github.insanusmokrassar.TelegramBotAPI.utils.extensions.toHtml
 import com.github.insanusmokrassar.TelegramBotAPI.utils.extensions.toMarkdown
 
 interface MessageEntity {
@@ -9,4 +10,7 @@ interface MessageEntity {
     
     val asMarkdownSource: String
         get() = sourceString.toMarkdown()
+
+    val asHtmlSource: String
+        get() = sourceString.toHtml()
 }