+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/TextLinkMessageEntity.kt
@@ -5,8 +5,7 @@ data class TextLinkMessageEntity(
     override val length: Int,
     override val sourceString: String,
     val url: String
-) : MessageEntity {
-    override val asMarkdownSource: String by lazy {
-        "[$sourceString]($url)"
-    }
-}
+) : LinkMessageEntity(
+    sourceString,
+    url
+)