+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/TextMessageEntity.kt
@@ -1,9 +1,14 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity
 
 abstract class TextMessageEntity : MessageEntity {
-    protected abstract val formatSymbol: String
+    protected abstract val markdownFormatSymbol: String
+    protected abstract val htmlFormatTagname: String
 
     override val asMarkdownSource: String by lazy {
-        "$formatSymbol$sourceString$formatSymbol"
+        "$markdownFormatSymbol$sourceString$markdownFormatSymbol"
+    }
+
+    override val asHtmlSource: String by lazy {
+        "<$htmlFormatTagname>$sourceString</$htmlFormatTagname>"
     }
 }