+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/URLMessageEntity.kt
@@ -4,11 +4,9 @@ data class URLMessageEntity(
     override val offset: Int,
     override val length: Int,
     override val sourceString: String
-) : MessageEntity {
-    val url: String
-        get() = sourceString
-
-    override val asMarkdownSource: String by lazy {
-        "[$sourceString]($url)"
-    }
+) : LinkMessageEntity(
+    sourceString,
+    sourceString
+) {
+    val url: String = link
 }